refactore tests
